UAST: split KotlinConverter into a separate file
